refactor: move typings and functions to packages/starlight/utils
